precompute nn ranking on full data
